Improve folder structure in filterBar
Also fix/simplify config propagation in Dashboard, unify the Hidden
versions of the components.

JIRA: RAIL-3383
