+++ libs/sdk-ui-dashboard/src/filterBar/DefaultFilterBar/DefaultFilterBarContainer.tsx
@@ -1,41 +1,14 @@
 // (C) 2021 GoodData Corporation
 import React, { useState } from "react";
 import Measure from "react-measure";
-import { FormattedMessage } from "react-intl";
 
-import { IntlWrapper } from "../localization/IntlWrapper";
+import { IntlWrapper } from "../../localization/IntlWrapper";
+import { ShowAllFiltersButton } from "./ShowAllFiltersButton";
 
 // TODO: this will probably need to be customizable so that custom filter components work
 const DEFAULT_FILTER_BAR_HEIGHT = 58;
 
-const ShowAllFiltersButton: React.FC<{
-    isVisible: boolean;
-    isFilterBarExpanded: boolean;
-    onToggle: (isExpanded: boolean) => void;
-}> = ({ isVisible, isFilterBarExpanded, onToggle }) => {
-    if (!isVisible) {
-        return null;
-    }
-
-    const icon = isFilterBarExpanded ? "gd-icon-chevron-up" : "gd-icon-chevron-down";
-
-    return (
-        <div className="show-all">
-            <button
-                className="button-filter-bar-show-all"
-                tabIndex={-1}
-                onClick={() => onToggle(!isFilterBarExpanded)}
-            >
-                <span className="gd-button-text gd-label">
-                    <FormattedMessage id={isFilterBarExpanded ? "filterBar.showLess" : "filterBar.showAll"} />
-                    <i className={`gd-button-icon gd-icon ${icon}`} />
-                </span>
-            </button>
-        </div>
-    );
-};
-
-const FilterBarContainerCore: React.FC = ({ children }) => {
+const DefaultFilterBarContainerCore: React.FC = ({ children }) => {
     const [filterBarHeight, setFilterBarHeight] = useState(DEFAULT_FILTER_BAR_HEIGHT);
     const [filterBarExpanded, setFilterBarExpanded] = useState(false);
 
@@ -77,10 +50,10 @@ const FilterBarContainerCore: React.FC = ({ children }) => {
 /**
  * @internal
  */
-export const FilterBarContainer: React.FC = ({ children }) => {
+export const DefaultFilterBarContainer: React.FC = ({ children }) => {
     return (
         <IntlWrapper>
-            <FilterBarContainerCore>{children}</FilterBarContainerCore>
+            <DefaultFilterBarContainerCore>{children}</DefaultFilterBarContainerCore>
         </IntlWrapper>
     );
 };
