+++ libs/sdk-ui-dashboard/src/filterBar/DefaultFilterBar/HiddenDashboardDateFilter.tsx
@@ -1,0 +1,14 @@
+// (C) 2021 GoodData Corporation
+import React from "react";
+
+import { IDashboardDateFilterCoreProps } from "../types";
+
+/**
+ * This implementation of dashboard date filter keeps the filter hidden out of sight. The attribute filter itself
+ * will still be in effect.
+ *
+ * @internal
+ */
+export const HiddenDashboardDateFilter: React.FC<IDashboardDateFilterCoreProps> = (_props) => {
+    return null;
+};
