+++ libs/sdk-ui-dashboard/src/filterBar/DefaultFilterBar/index.ts
@@ -1,0 +1,7 @@
+// (C) 2021 GoodData Corporation
+export { DefaultFilterBar } from "./DefaultFilterBar";
+export { DefaultDashboardAttributeFilter } from "./DefaultDashboardAttributeFilter";
+export { DefaultDashboardDateFilter } from "./DefaultDashboardDateFilter";
+
+export { HiddenDashboardAttributeFilter } from "./HiddenDashboardAttributeFilter";
+export { HiddenDashboardDateFilter } from "./HiddenDashboardDateFilter";
