+++ libs/sdk-ui-dashboard/src/topBar/DefaultTopBar/HiddenButtonBar.tsx
@@ -1,0 +1,13 @@
+// (C) 2021 GoodData Corporation
+import React from "react";
+
+import { IButtonBarCoreProps } from "../types";
+
+/**
+ * This implementation of button bar keeps it out of sight and effectively disables it.
+ *
+ * @internal
+ */
+export const HiddenButtonBar: React.FC<IButtonBarCoreProps> = (_props) => {
+    return null;
+};
