+++ libs/sdk-ui-dashboard/src/topBar/DefaultTopBar/HiddenTitle.tsx
@@ -1,0 +1,13 @@
+// (C) 2021 GoodData Corporation
+import React from "react";
+
+import { ITitleCoreProps } from "../types";
+
+/**
+ * This implementation of title keeps it out of sight.
+ *
+ * @internal
+ */
+export const HiddenTitle: React.FC<ITitleCoreProps> = (_props) => {
+    return null;
+};
