+++ libs/sdk-ui-dashboard/src/topBar/DefaultTopBar/index.tsx
@@ -3,3 +3,7 @@ export { DefaultTopBar } from "./DefaultTopBar";
 export { DefaultButtonBar } from "./DefaultButtonBar";
 export { DefaultMenuButton } from "./DefaultMenuButton";
 export { DefaultTitle } from "./DefaultTitle";
+
+export { HiddenButtonBar } from "./HiddenButtonBar";
+export { HiddenMenuButton } from "./HiddenMenuButton";
+export { HiddenTitle } from "./HiddenTitle";
