+++ libs/sdk-ui-dashboard/src/topBar/HiddenTopBar.tsx
@@ -1,10 +1,13 @@
 // (C) 2021 GoodData Corporation
 import React from "react";
+
 import { ITopBarCoreProps } from "./types";
 
 /**
+ * This implementation of top bar keeps it out of sight.
+ *
  * @internal
  */
-export const NoTopBar: React.FC<ITopBarCoreProps> = () => {
+export const HiddenTopBar: React.FC<ITopBarCoreProps> = (_props) => {
     return null;
 };
